Support simulated location for testable targets
