+++ Sources/TuistGraph/Models/TestableTarget.swift
@@ -1,5 +1,4 @@
 import Foundation
-import TSCBasic
 
 /// Testable target describe target and tests information.
 public struct TestableTarget: Equatable, Hashable, Codable {
@@ -11,16 +10,20 @@ public struct TestableTarget: Equatable, Hashable, Codable {
     public let isParallelizable: Bool
     /// Execute tests in random order.
     public let isRandomExecutionOrdering: Bool
+    /// A simulated location used when testing this test target.
+    public let simulatedLocation: SimulatedLocation?
 
     public init(
         target: TargetReference,
         skipped: Bool = false,
         parallelizable: Bool = false,
-        randomExecutionOrdering: Bool = false
+        randomExecutionOrdering: Bool = false,
+        simulatedLocation: SimulatedLocation? = nil
     ) {
         self.target = target
         isSkipped = skipped
         isParallelizable = parallelizable
         isRandomExecutionOrdering = randomExecutionOrdering
+        self.simulatedLocation = simulatedLocation
     }
 }
